requirements-cloudrun.txt
# Core dependencies for Google Cloud Run
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI and ML dependencies
anthropic==0.40.0
scikit-learn==1.3.2
numpy==1.24.3
Pillow==10.1.0

# PDF processing
PyMuPDF==1.23.8

# Document AI
google-cloud-documentai==2.20.1
google-auth==2.17.3

# Google Cloud Storage
google-cloud-storage==2.10.0
google-cloud-run==0.10.0

# Database and storage
redis==5.0.1

# Utilities
psutil==5.9.6
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Monitoring and logging
structlog==23.2.0
